Agent: various compiler warning fixes
Includes fixes for VxWorks-specific logic, (missing function
declarations, etc.) and in particular, typing for 64-bit platforms.
Also fixes diab's reporting of unnecessary variable setting, and removes
a few simple conditionals which always evaluate to true.

+++ agent/tcf/framework/trace.c
@@ -116,7 +116,7 @@ int parse_trace_mode(const char * mode, int * result) {
     for(;;) {
         if (*mode >= '0' && *mode <= '9') {
             char * endptr;
-            *result |= strtoul(mode, &endptr, 0);
+            *result |= (int) strtoul(mode, &endptr, 0);
             mode = endptr;
         }
         else {
